requirements.txt
pandas==2.3.3
python-dotenv==1.2.2
requests==2.32.5
SQLAlchemy==2.0.48
streamlit==1.55.0
tqdm==4.67.3
